@@ -14,6 +14,6 @@ class OnStartRecording extends RemoConnectionEvent {}
 /// Used to select a specific device among the ones discovered.
 class OnSelectDevice extends RemoConnectionEvent {
   OnSelectDevice(this.deviceName, this.deviceAddress);
-  String deviceAddress;
-  String deviceName;
+  String? deviceAddress;
+  String? deviceName;
 }
